refactor logic into service, so clean
